--- src/Components/Header.jsx
@@ -1,3 +1,0 @@
-export default function Header() {
-  return;
-}
